+++ libs/src/ui/element/Radio/RadioGroup.vue
@@ -1,17 +1,17 @@
 <script setup>
 // 定義 Props
 const props = defineProps({
-  // --  基礎接口 -- //
-  customClass: {
-    type: String,
-    default: '',
-  },
-  //--  layout 接口 -- //
-  direction: {
-    type: String,
-    validator: (value) => ['row', 'column'].includes(value),
-    default: 'row',
-  },
+	//--  layout 接口 -- //
+	direction: {
+		type: String,
+		validator: (value) => ['row', 'column'].includes(value),
+		default: 'row',
+	},
+	// --  基礎接口 -- //
+	customClass: {
+		type: String,
+		default: '',
+	},
 })
 </script>
